Added logging/debug mechanism to fastcgi++
This adds support for a simplified logging mechanism to fastcgi++. It allows
users to add logging code that can be ignored by the compiler based on a log
level supplied at compile time.

The log destination can either be standard error or a custom log file. Log lines
can have timestamps associated with them or disabled.